spenzy-document-service/requirements.txt
grpcio>=1.68.0
grpcio-tools>=1.68.0
grpcio-reflection>=1.68.0
python-jose==3.3.0
pyjwt==2.10.1
cryptography==44.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pdf2image==1.16.3
httpx==0.27.0
python-magic>=0.4.27
python-keycloak>=3.0.0
-e ../spenzy-common